jetstream: set tcp timeouts
for the moment it's arbitrarily hard-coded at 8s for connect, and 4s for both read and write.

i'm not really sure what the problem would be for doing it this way, but according to the tungstenite author, [i'm probably doing it wrong](https://github.com/snapview/tungstenite-rs/issues/252#issuecomment-984144901)

mayyyybe using tungstunite-rs is too low level and i really should embrace #[tokio:main]. probably. but i'm going to hold out a minute longer.