first attempt to re-adapt for server
laksdjflkasjdflkajsdflkjasldfkja

can't use generics in server context for dropshot so everything is pain

- we can use a dyn thing to get the store reader into the context
- but then we can't be clone, so
- we cannot spawn_blocking to wrap the db calls because
- dropshot only gives us a borrow of the context and
- tokio spawn_blocking needs 'static

we *can* use block_in_place, but this doesn't feel particularly good.

we could hard-code the server to use the fjall store, but this feels worse.

it might be possible to use async-trait and push the spawning down to the store internals, who can clone the necessary things to become 'static i think? will try this next.